backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-multipart
pillow
google-generativeai
